backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
supabase==2.10.0
requests==2.32.3
langchain==0.1.0
langchain-text-splitters==0.0.1
langchain-ollama==0.3.10
groq==0.32.0 